# docs/api/openapi.public.yaml
components:
    schemas:
...
        github_com_trakrf_platform_backend_internal_models_location.CreateLocationWithIdentifiersRequest:
            properties:
                description:
                    example: Main warehouse location
                    maxLength: 1024
                    type: string
                identifier:
                    example: wh1
                    maxLength: 255
                    minLength: 1
                    type: string
                identifiers:
                    items:
                        $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_shared.TagIdentifierRequest'
                    type: array
                is_active:
                    example: true
                    type: boolean
                name:
                    example: Warehouse 1
                    maxLength: 255
                    minLength: 1
                    type: string
                parent_location_id:
                    example: 1
                    minimum: 1
                    type: integer
                valid_from:
                    example: "2025-12-14"
                    type: string
                valid_to:
                    example: "2026-12-14"
                    type: string
            required:
                - identifier
                - name
            type: object
...
        github_com_trakrf_platform_backend_internal_models_shared.TagIdentifierRequest:
            properties:
                type:
                    enum:
                        - rfid
                        - ble
                        - barcode
                    example: rfid
                    type: string
                    x-extensible-enum: "true"
                value:
                    maxLength: 255
                    minLength: 1
                    type: string
            required:
                - value
            type: object
    securitySchemes:
        APIKey:
            bearerFormat: JWT
            description: 'TrakRF API key (JWT). Format: "Bearer <jwt>". Mint keys from the API Keys section of your TrakRF account (key-management UI is tracked by TRA-393).'
            scheme: bearer
            type: http
        BearerAuth:
            description: Session JWT for internal endpoints (platform frontend uses this).
            in: header
            name: Authorization
            type: apiKey
info:
    contact:
        email: support@trakrf.id
        name: TrakRF Support
    description: TrakRF public REST API. See docs.trakrf.id/api for the customer-facing reference.
    license:
        name: Business Source License 1.1
        url: https://github.com/trakrf/platform/blob/main/LICENSE
    title: TrakRF API
    version: v1
openapi: 3.0.3
paths:
...
    /api/v1/locations:
        get:
            parameters:
                - description: max 200
                  in: query
                  name: limit
                  schema:
                    type: integer
                - description: pagination offset
                  in: query
                  name: offset
                  schema:
                    type: integer
                - description: filter by parent identifier (may repeat)
                  in: query
                  name: parent
                  schema:
                    type: string
                - description: filter by active flag
                  in: query
                  name: is_active
                  schema:
                    type: boolean
                - description: fuzzy search on name, identifier, description
                  in: query
                  name: q
                  schema:
                    type: string
                - description: comma-separated, prefix '-' for DESC
                  in: query
                  name: sort
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: OK
            security:
                - BearerAuth: []
            summary: List locations
            tags:
                - locations
        post:
            description: |-
                Create a new location in the hierarchy, optionally with one or more tag identifiers.
                Set ParentLocationID to nest the location under an existing parent. The Location response header contains the canonical URL.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_location.CreateLocationWithIdentifiersRequest'
                description: Location to create with optional identifiers
                required: true
                x-originalParamName: request
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: 'data: location.LocationView'
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: bad_request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: unauthorized
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: forbidden
                "429":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: rate_limited
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: internal_error
            security:
                - APIKey:
                    - locations:write
            summary: Create a location
            tags:
                - locations
...
    /api/v1/locations/{id}/ancestors:
        get:
            description: Return all ancestor locations from the root of the hierarchy down to the immediate parent of the specified location.
            parameters:
                - description: Location ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "202":
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: 'data: []location.Location'
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: bad_request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: unauthorized
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: forbidden
                "429":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: rate_limited
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: internal_error
            security:
                - APIKey:
                    - locations:read
            summary: List location ancestors
            tags:
                - locations
    /api/v1/locations/{id}/children:
        get:
            description: Return the immediate child locations of the specified location (one level deep only).
            parameters:
                - description: Location ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "202":
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: 'data: []location.Location'
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: bad_request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: unauthorized
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: forbidden
                "429":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: rate_limited
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github_com_trakrf_platform_backend_internal_models_errors.ErrorResponse'
                    description: internal_error
            security:
                - APIKey:
                    - locations:read
            summary: List location children
            tags:
                - locations
...
servers:
    - description: Production
      url: https://app.trakrf.id
    - description: Preview (per-PR deploys)
      url: https://app.preview.trakrf.id